Update spell-crafter-mainnet-workflow.md
Minor change.

From All to Crafter. If we put ALL, then it is no one's responsibility. so I prefer that the Crafter starts the retro.

From (if needed) to Always: Retro's should always happen, if not, starting a retro can be perceived as finger pointing. Also, if the retro is mandatory, it forces ppl to raise topics before they become a bigger issue that would create discomfort. Having a retro as a habit generates a place for psychological safety where members can share anything, even if it is not a big deal. If there is nothing to discuss in the retro, then a simple one-liner by the Crafter: "Retro:Nothing to discuss" is enough, the rest can thumbs up or share something from their side.